+++ src/libMediator/Mediator.cpp
@@ -21,6 +21,7 @@
 #include "common/Constants.h"
 #include "libCrypto/Sha2.h"
 #include "libEth/Filters.h"
+#include "libServer/DedicatedWebsocketServer.h"
 #include "libServer/GetWorkServer.h"
 #include "libUtils/CommonUtils.h"
 #include "libUtils/DataConversion.h"
@@ -59,7 +60,8 @@ Mediator::Mediator(const PairOfKey& key, const Peer& peer)
               TX_DISTRIBUTE_TIME_IN_MS +
               (DS_ANNOUNCEMENT_DELAY_IN_MS + SHARD_ANNOUNCEMENT_DELAY_IN_MS)) /
           1000),
-      m_filtersAPICache(evmproj::filters::APICache::Create()) {
+      m_filtersAPICache(evmproj::filters::APICache::Create()),
+      m_websocketServer(rpc::DedicatedWebsocketServer::Create()) {
   SetupLogLevel();
 }
