+++ src/libMediator/Mediator.h
@@ -36,6 +36,10 @@ class APICache;
 }
 }  // namespace evmproj
 
+namespace rpc {
+class DedicatedWebsocketServer;
+}
+
 /// A mediator class for providing access to global members.
 class Mediator {
  public:
@@ -120,6 +124,9 @@ class Mediator {
 
   /// Filters API cache
   std::shared_ptr<evmproj::filters::APICache> m_filtersAPICache;
+
+  /// Dedicated (old) websocket server
+  std::shared_ptr<rpc::DedicatedWebsocketServer> m_websocketServer;
 
   /// Constructor.
   Mediator(const PairOfKey& key, const Peer& peer);
