+++ src/libServer/WebsocketServerImpl.cpp
@@ -23,6 +23,24 @@
 #include "libUtils/Logger.h"
 
 namespace rpc {
+
+/// Creates WS backend w/o threadpool: for Scilla WS server
+std::shared_ptr<WebsocketServerBackend> WebsocketServerBackend::Create(
+    AsioCtx& asio) {
+  return std::make_shared<ws::WebsocketServerImpl>(
+      asio, std::shared_ptr<APIThreadPool>{});
+}
+
+/// Creates WS backend w/threadpool: for jsonrpc API server
+std::shared_ptr<WebsocketServerBackend>
+WebsocketServerBackend::CreateWithThreadPool(
+    AsioCtx& asio, std::shared_ptr<APIThreadPool> threadPool) {
+  if (!threadPool) {
+    throw std::runtime_error("Thread pool expected");
+  }
+  return std::make_shared<ws::WebsocketServerImpl>(asio, std::move(threadPool));
+}
+
 namespace ws {
 
 using CloseReason = websocket::close_code;
@@ -39,22 +57,27 @@ class Connection : public std::enable_shared_from_this<Connection> {
       : m_owner(std::move(owner)),
         m_id(id),
         m_from(std::move(from)),
-        m_stream(std::move(socket)) {}
-
-  void WebsocketAccept(HttpRequest&& req) {
+        m_stream(std::move(socket)) {
     websocket::stream_base::timeout opt{
         std::chrono::seconds(30),  // handshake timeout
         std::chrono::seconds(10),  // idle timeout
         true                       // pings
     };
     m_stream.set_option(opt);
-
     m_stream.set_option(
         websocket::stream_base::decorator([](websocket::response_type& res) {
-          res.set(http::field::server, ".");  // TODO server string for http
+          res.set(http::field::server, "zilliqa");
         }));
+  }
+
+  void WebsocketAccept(HttpRequest&& req) {
     m_stream.async_accept(
         std::move(req),
+        beast::bind_front_handler(&Connection::OnAccept, shared_from_this()));
+  }
+
+  void WebsocketAccept() {
+    m_stream.async_accept(
         beast::bind_front_handler(&Connection::OnAccept, shared_from_this()));
   }
 
@@ -214,13 +237,22 @@ void WebsocketServerImpl::CloseAll() {
 
 void WebsocketServerImpl::NewConnection(std::string&& from, Socket&& socket,
                                         HttpRequest&& req) {
+  CreateNewConnection(std::move(from), std::move(socket))
+      ->WebsocketAccept(std::move(req));
+}
+
+void WebsocketServerImpl::NewConnection(std::string&& from, Socket&& socket) {
+  CreateNewConnection(std::move(from), std::move(socket))->WebsocketAccept();
+}
+
+std::shared_ptr<Connection> WebsocketServerImpl::CreateNewConnection(
+    std::string&& from, Socket&& socket) {
   auto conn = std::make_shared<Connection>(weak_from_this(), ++m_counter,
                                            std::move(from), std::move(socket));
-  conn->WebsocketAccept(std::move(req));
-  m_connections[m_counter] = std::move(conn);
-
+  m_connections[m_counter] = conn;
   LOG_GENERAL(INFO, "WS connection #" << m_counter << " from " << from
                                       << ", total=" << m_connections.size());
+  return conn;
 }
 
 bool WebsocketServerImpl::MessageFromConnection(ConnectionId id,
@@ -263,7 +295,7 @@ bool WebsocketServerImpl::MessageFromConnection(ConnectionId id,
     return false;
   }
 
-  if (unknownMethodFound) {
+  if (unknownMethodFound && m_threadPool) {
     // forward msg to the thread pool
     // TODO double json parsing: to be fixed after project dependencies change
     if (!m_threadPool->PushRequest(id, true, from, std::move(msg))) {
