+++ src/libZilliqa/Zilliqa.cpp
@@ -32,9 +32,9 @@
 #include "libNetwork/Guard.h"
 #include "libRemoteStorageDB/RemoteStorageDB.h"
 #include "libServer/APIServer.h"
+#include "libServer/DedicatedWebsocketServer.h"
 #include "libServer/GetWorkServer.h"
 #include "libServer/LocalAPIServer.h"
-#include "libServer/WebsocketServer.h"
 #include "libUtils/DetachedFunction.h"
 #include "libUtils/Logger.h"
 #include "libUtils/SetThreadName.h"
@@ -444,7 +444,7 @@ Zilliqa::Zilliqa(const PairOfKey& key, const Peer& peer, SyncType syncType,
       }
 
       if (ENABLE_WEBSOCKET) {
-        (void)WebsocketServer::GetInstance();
+        m_mediator.m_websocketServer->Start();
       }
 
       if (m_lookupServer == nullptr) {
@@ -539,7 +539,10 @@ Zilliqa::Zilliqa(const PairOfKey& key, const Peer& peer, SyncType syncType,
   DetachedFunction(1, func);
 }
 
-Zilliqa::~Zilliqa() { m_msgQueue.stop(); }
+Zilliqa::~Zilliqa() {
+  m_msgQueue.stop();
+  m_mediator.m_websocketServer->Stop();
+}
 
 void Zilliqa::Dispatch(Zilliqa::Msg message) {
   LOG_MARKER();
